Start adding data plane tests for MMU tuning
Summary:
The idea here is to test for
i) MMU tuning parameters programming and warmboot
ii) Effect on queue buffer utilization from this tuning

i) is accomplished by the fact that we ask the HwSwitch to
program these parameters. So barring the pathological case
where HwSwitch is completely ignoring this config input test
verifies the programming and warmboots across it

ii) is harder.  Even with line rate traffic, its hard to create contention in MMU
buffer, and thus we get roughly similar watermark values. To create
contention, we would need to create enough bursts of traffic such that
the MMU is forced to throttle back the low pri queue. So as a compromise
we do the next best thing - compare that the high pri queue watermark
is at least as high as the low pri one. This lets us at least verify
the config application with MMU tuning and warm boot. Although to be
fair, in absence of precise data plane side effects we can't truly
conclude on the tuning effects.

Reviewed By: shri-khare

Differential Revision: D24143223

fbshipit-source-id: 6a403ca3ecf49c60c58fb816006ad658bfa5ce6f