Add FBOSS_ALERT(BGP) tags to monitor BGP sessions
Summary:
One of the DC Cohort Goals is to add tags to high signal log messages in
order to add them to scuba for Visibility.  See https://fb.quip.com/UPjcAghowIan.
In this diff we add tags for alerts to do with sessions BGP sessions coming
up and going down, including tracking on when EOR is received and final
convergence.  More tags to be added in other commits.

Production Rollout Plan:
No special considerations.  This is not dependent on pervious binary
version, or on any config parameters.

Reviewed By: rsunkad

Differential Revision: D24214303

fbshipit-source-id: d1d88eea6306d5c4a39fcc57deddcd069dbcd06e