Handle QoS policy name change
Summary:
Agent can handle changes to qos policy scheme but cannot handle qos policy name
change itself. This is because the implementation tries to query the changed
(new) name in BcmSwitch's local data structures, and if it finds nothing, it
throws an error.

When D24320929 changed the qos policy name as part of removing ECN queue 2,
this caused canary failures S212947.

This is temporarily mitigated by D24367813 that reverts the qos policy name,
but:

- we need to handle qos policy name changes in agent code,
- 672 switches are already updated to new name: P145907321
  which means that while D24367813 prevents the problem from spreading, it puts
  those 672 switches in the risk to hit the agent crash unless the issue is
  fixed.

This patch fixes the agent issue by simply removing and readding the QosPolicy
on name change. We could consider optimizing it to handle name change only case
as follow up, but this providens the necessary relief for those 672 switches
and beyond.

Reviewed By: jasmeetbagga

Differential Revision: D24371354

fbshipit-source-id: 50e2597d8fd4a756883be29633b11bddf5a7fbd4