deep copy of fdb event notifications
Summary:
When FDB event is received, making a copy of the struct alone is not sufficient. The attribute in sai_fdb_event_notification_data_t is a pointer which needs to be copied too. If not, the fdb eventbase thread will be accessing some invalid memory.

This can be achived in several ways:

1) Deep copy of sai_fdb_event_notification_data_t. This includes allocating memory for sai_attribute_t in the fdb event data and freeing them in the fdb bottom half thread.

2) Create a temporary struct FdbEventNotificationData which is a subset of sai_fdb_event_notification_data_t. This will be used to store only the necessary attributes and pass it on to fdb bottom half thread.

3) Avoid creating temporary structs and create L2Entry which can be sent to sw switch. The problem with this approach is we need to hold the sai switch mutex in the callback thread. Also, this can cause deadlock when we used the bridge port id to fetch the port id from the SDK.

I am going with option 2.

Reviewed By: jasmeetbagga

Differential Revision: D24383073

fbshipit-source-id: c63faa846fa9dfe1c6e24a4c44dc540369dbf86d