Set TC for Traffic to CPU
Summary:
This is towards supporting ACLs in SAI implementation.

Set TC for traffic to CPU.

This works because:

By default BCM maps TC i to Queue for i in [0, 9].
Tajo claims to map TC i to Queue i by default as well.

Subsequent stacked diff will explicitly set this mapping and associate
that with the CPU port.

Differential Revision: D24414371

fbshipit-source-id: ed13fc5c3fbb883e1b227905e50466932bbfe8bd