Reorder AclTable ActionTypeList
Summary:
This is towards supporting ACLs in SAI implementation.

In the current implementation, Tajo getter for ActionTypeList hard codes a
list. The list elements are identical to what FBOSS sets but are in different
order. Since they are stored in unordered data structure (vector), the
comparison fails and we end up creating yet another ACL table post wb.

This is being tracked and would be fixed in Tajo. In the meantime, match the
order to unblock tests.

Differential Revision: D24450511

fbshipit-source-id: bcd5ae3245ba963c22d8eae857471960f250bc84