Aggregate member port incorrectly added to enabled state
Summary: A LAG member port can be incorrectly enabled during lacp syncing. If a member port is in detached or waiting state and we receive an LACP PDU with matching information, the port will be marked as enabled even if mux is in detached or waiting state. This is a regression from the changes of D24055517 (https://github.com/facebook/fboss/commit/3e5067827aa39c82a8ca364bdc46e83ad68c49fe) and was observed while testing warmboot on DNE testbed.

Reviewed By: rsunkad

Differential Revision: D24497012

fbshipit-source-id: b2bf5c7b050feafc8d286720260516edd3d52dfa