Refactor BcmStatUpdater get acl stat
Summary:
The current BcmStatUpdater keeps the bcm_field_stat_get() inside its own .cpp, which is fine for non-TH4 platforms.
But TH4 needs to use FlexCounter to implement ingress field processor stats(will be implemented in the next couple diffs).
Besides, we already have a BcmAclStat class, which will be the perfect place to abstract a getStats() function, and inside its implementation,
it can pick to use whether bcm_field_stat_get() or FlexCounter.
So BcmStatUpdater can just call BcmAclStat::getStats() to collect the stats of TrafficCounter ACL.

In order to do that, I also had to refactor how BcmStatUpdater store the acl counter.
Currently, it uses AclCounterDescriptor as a key, which has BcmAclStatHandle and CounterType.
This might not be enough to look for BcmAclStat because BcmSwitch::BcmAclTable is actually using the counter name to store BcmAclStat.
Therefore, I changed the counter map in BcmStatUpdater a little bit to:
```
   /* ACL stats */
  // The string is counter name
  std::queue<std::string> toBeRemovedAclStats_;
  std::queue<std::pair<std::string, cfg::CounterType>> toBeAddedAclStats_;
  // Outer-map key: counter name
  // Inner-map key: counter type, value: MonotonicCounter
  folly::Synchronized<std::unordered_map<
      std::string,
      std::unordered_map<cfg::CounterType, std::unique_ptr<MonotonicCounter>>>>
      aclStats_;

```

Reviewed By: jasmeetbagga

Differential Revision: D24011594

fbshipit-source-id: 90da6e3167d8640b6f8706912251bdb3f8ede68b