Explicitly enable bcmCosqControlEgressUCSharedDynamicEnable before programming alpha for the egress queues
Summary:
Alpha programming for the egress queues require 'bcmCosqControlEgressUC(MC)SharedDynamicEnable' to be set. In all HW platforms in current mode, SDK init ensures that this register is set. So explicitly setting/checking this is a no-op. However its not set in the following cases:
(1) SIM
(2) mmu_lossless=1

This change covers set/get aspects of the same.

Reviewed By: jasmeetbagga

Differential Revision: D24183691

fbshipit-source-id: 82fdf09d4ce1110856178057a0d6b3f36c2d7ea6