Fix LACP slow start after warm boot
Summary:
If LACP was up before warm boot, we restore the previous state. If LACP was down before warm boot, we start LACP state machine as normal. In the second case LACP was not converging because it expects a port up event before accepting PDUs and there is no port up event in case of warm boot init.

Fix is to call lacp port up event after warm boot if port is up. Also converted some of the logs to DBG2

Reviewed By: rsunkad

Differential Revision: D24559768

fbshipit-source-id: efd007db37793b4c6c9f225a5f1bbb4530cd2883