Fix remote side parameters used by lacp selection logic on warm boot restore
Summary:
We restore LACP state machines after warm boot in a specific order. selector was using default values for remote side parameters because of the order in which restoration was happening. The functionality was working but was taking an unexpected path.

Fixed the order in which restoration happens so that selector uses right parameter. Also modified selector restore to do directly change to final state as calling select() here can take the link to standby initially.

Reviewed By: jasmeetbagga

Differential Revision: D24559781

fbshipit-source-id: b1e175e67df81648c172a66ad604523d4da14f15