Fix scheduler obj in use error on scheduler config change
Summary:
With D24422314 (https://github.com/facebook/fboss/commit/81a77d1914690f74da38da1f2023e5ac56c7f06b) we made 2Q->Olympic qos tests
more representative by adding queue configs to the mix.
This exposed a bug in the way we were setting scheduler profiles,
consider
- We would update queueHandle->scheduler shared_ptr reference
before updating the SchedulerProfileId attr. This meant that
if this queue already had a scheduler and happened to be the
last queue, we would try to delete the scheduler before updating
the queue reference. This lead to obj in use error
- If scheduler were ever set to null (not the case in our code,
since we default to WRR with weight 1), we would would not go
and update the queue scheduler to null obj id

Differential Revision: D24585226

fbshipit-source-id: dd8157bad686a25f88cfc40331343812dcb153f0