Add turbo controller use case to benchmark
Summary:
Enhanced turbo benchmark to use UCMP. For turbo controller use case, ucmp will be enabled on the mesh links. The test uses 1:3 weight distribution to create ECMP with 32 ip and 96 labelled paths. The results are very similar to ECMP case with 64 paths which is expected because for ucmp we don't create extra label paths and instead repeat the paths in the ecmp structure.

Also made a change in netcastle benchmark script to pass ecmp_width command line arg

Reviewed By: cenzhao, phshaikh, jhshi

Differential Revision: D24342956

fbshipit-source-id: ffa5c73c68a6ab20b79b3603ade75bd8796eae15