Add LACP unit test for link state changes
Summary: The purpose of this test is to catch the scenario of S213793. The test creates an Aggregateport with 2 member links between DU and UU with minlink set to 2. One of the member gets a port down event. This causes both links to change to fwding disabled state. The second port is in detached state but is physically up. If it receives a matching PDU in this state with partner state changed, it should not change to fwding up state.

Reviewed By: jasmeetbagga

Differential Revision: D24584648

fbshipit-source-id: 4c61301288c5865374722ab4663f22476e340153