Helper method in SaiStoreTest for creating SaiObjects
Summary:
As titled. This is working towards making
SaiObject constructors private and accessible only
to SaiObjectStore and tests. We want SaiObjects to
be created through the store so that they can be retrieved
correctly on WB. There have been a handful of bugs where
we ended up bypassing the store and creating standalone
SaiObjects. Lets get compiler to help us here

Differential Revision: D24721926

fbshipit-source-id: 61013a20366e500dee9c692df1d2b25744c33089