Add a refOrInsert API to RefMap
Summary:
In some use cases we want to pass in a already
constructed value for refmap to insert. A example of this is
for Sai Objects where we want to construct them *only* from
SaiStore. SaiStore in turn then manages these objects in a ref
map. To that end we need a insert and not just emplace, since
that latter would want to construct SaiObjects

TODO:
Briefly making public the constructors of SaiObject since SaiObjectStore
still uses refOrEmplace APIs. And to allow that we would have to
open up access to std::make_unique. Making private constructors worthless.
Next diffs will update SaiObjectStore and bring private constructors
back

Reviewed By: phshaikh

Differential Revision: D24742504

fbshipit-source-id: ef0d2a4c51c2273419c5ed7b0d22208658196123