Support get number of acl stat in acl group using FlexCounter
Summary:
In our bcm_test, we use fpGroupNumAclStatEntries() from BcmFieldProcessorUtils.h to get the number of acl stat in specific fpGroup.
This function has been called in multiple test cases in HwAclStatTest.cpp,
and it uses `bcm_field_group_status_get()`.

However, since we have to use FlexCounter in TH4, the above function will only return 0 acl stat for the fp group.
So to support such logic in TH4, we need to implement the FlexCounter way to calculate the acl stat belongs to a specific group.

Reviewed By: jasmeetbagga

Differential Revision: D24612494

fbshipit-source-id: 56c90ea13150f43f50a86ccb3e665561d05578b3