Support get acl stat by acl during WB for FlexCounter
Summary:
During warmboot, we need to use populateAclStats() to get the acl stat handle for a specific acl entry.
Once again, because we have to use FlexCounter for TH4, we also need to change the implementation of
getting the acl stat handle back.

Currently the logic is in BcmWarmbootCache and I think it might be more helpful to keep all the AclStat logic just in BcmAclStat.

Reviewed By: jasmeetbagga

Differential Revision: D24641600

fbshipit-source-id: b17ce5ad0ce35640ad98b1553dabbbf70cc87f72