Adjust disabled port profileID and speed based on VCO difference
Summary:
Currently we use 200G as a default port speed in the agent.config template for TH4 platforms,
but we still need to support 100G for some fabric doesn't have 200G support.

And in TH4, each Blackhawk7 core can only use one VCO frequency. And 200G with RS544 needs 26GHz VCO;
while 100G with RS528 needs 25GHz VCO.

Therefore, if one of the ports from the same core needs to use 100G, and the others use 200G,
SDK will throw error when we try to use FlexPort feature.

Therefore, we need to add an override implementation for fill_interfaces() of FbossFujiFswAgentConfigCompiler.
- First call the super class fill_interfaces() to make sure port state/ speed/ profileID/ fec are already updated based on source of truth.
- And then we check whether all the enabled ports are using the same VCO so that we can stop coop to generate the wrong config which can crash agent
- And then if there is *NOT* disabled port or the disabled ports are using the exact same VCO as enabled ports, then we're good
- Otherwise, we will update these disabled ports to use the exact same port profileID/ speed/ fec so that we make disabled ports to use the same VCO as enabled ports.

Since we're only changing the disabled ports, so it's quite safe to use this logic.

Differential Revision: D24775632

fbshipit-source-id: 9881959b143137308b0e41a77f9fa1e9419c3192