direct implement and clean up unused code
Summary:
This is a prototype of direct implementation of the transforms without conversion of the original data to auxiliary `ModelData` (as opposed to D24065135).
This diff is to illustrate the less heavy-weighted implementation of the idea compared to previous proposal, although there are some unimplemented parts and not the best quality of the code.

transforms implemented for port stats:
* RATE
* BYTES_TO_BITS (* 8)

Reviewed By: jasmeetbagga

Differential Revision: D24481856

fbshipit-source-id: 35994153c638111b6ed9a1e4d392b8c0c7bc79f4