deprecate getting list of supported profiles
Summary: We want to change the way we're getting supported profiles to be able to specify pim number. This is because on new platforms, a profileID can map to different profile configs, depending on which LC type. To help support this we can remove all cases of getSupportedProfiles, since we won't have a simple mapping anymore, and instead only rely on PlatformMapping::getPortProfileConfig to get a profile

Reviewed By: birdsoup

Differential Revision: D24808349

fbshipit-source-id: 7b38035757c55a527cb72bfda83167d94f3b91ab