Use cached vco frequency set in BcmPortGroup
Summary:
I noticed on TH4, when we programmed two 200G ports from the same group to 100G,
the same BcmPortGroup::reconfigureIfNeeded() has been called twice, even though each
time we program the same settings of these two ports. (P149585860)

The reason is when reconfigureIfNeeded() call needsVCOChange(), it uses the same old state to compare with
the new state. Since BcmSwitch::reconfigurePortGroups() will iterate all the ports one by one, the two ports from
the same group will be called sequentially. But we only update the SwitchState after all the settings are programmed
to the hardware.

Therefore, we can't just use old SwitchState to compare the VCO frequencies. What we can do is to mimic what we did
for BcmPortGroup::LaneMode to cache what we have programmed. And then use the cached set to compare with the new
SwitchState. In such way, if we already programmed the BcmPortGroup, then it won't program it again for the second port.

Differential Revision: D24945388

fbshipit-source-id: 6c2874093634e2bf5e8c41a8c04dddfaad49349b