Use appropriate max egress pool size for TH
Summary:
Not all of the MMU buffer is available for allocation
for the egress pool. BRCM chips on bootup reserve bytes for
CPU port and lb port queues. These are then not available for
use by the shared pool. Setting the shared pool to MMU buffer
size bytes messes up the shared pool size calculations and will
lead to silent drops in MMU.

TODO
- BRCM will provide a extension attr to get this value, once this
is available, we can assert/test for correctness.

Differential Revision: D25115935

fbshipit-source-id: 1c4da6ab685900677e45868410fdaf828a12bbe3