dst = src in _fill when dst and src are of same type
Summary:
As titled. We were missing the simple assignment
when types were the same. This bit us in getAttr for Sai
extension attributes (UTs stacked).
TODO: Investigate why we were getting away with this for
SaiAttribute. Possibly because we use WrappedSaiType/DuplicateTypeFixer
in that class for DataType vs using T directly

Reviewed By: phshaikh

Differential Revision: D25171972

fbshipit-source-id: 5226ad2344f4cfbd500b41cb22490470083eea4c