Assert for max buffer pool size on cold boot
Summary:
For some ASICs its crucial to start with the right
buffer pool size when creating this on cold boot. Defaulting
to MMU size will cause pre reserved bytes for not to be accounted
for. For such ASICs we look at the register value to get the starting
value. To ensure this is also correct in the future assert so
during cold boot *before* configuring the buffer pool.

Unfortunately we can't use this ext attribute as a size to set for
on pool creation. Reason for this is that once we configure a pool
this value will be updated. Then when we allot reserved bytes
to queues that value will be subtracted from. So on WB, we will come
back with a smaller value, and end up using that as our starting value.
Reserved bytes will now subtract from that. Continously decreasing the
size of available buffer on each warm boot. Hence use a fixed size and
assert that we are using the right value for a chip.

Reviewed By: phshaikh

Differential Revision: D25172344

fbshipit-source-id: 95f302e39213d2f421c46fbfa62a91af2a2ed80e