Add run_bcm_test.sh to remove setup overhead
Summary:
Due to the way setup_fboss_env is set up, running bcm_test programmatically requires navigating to $FBOSS directory, running "source bin/setup_fboss_env", and then run the actual bcm_test itself. Since children shells do not export any variables to the parent, we are having to repeat above steps each time we list tests / run a test.

To remove part of the overhead, adding a file that sets up variables and passes arguments to bcm_test. This should make it easier for both humans and scripts as this will remove the need to issue multiple commands to run the binary.

Reviewed By: rsunkad

Differential Revision: D25167306

fbshipit-source-id: 3bc312959d0cb1aa0d8146a073479d279f7191b4