Support TH4 queue default config
Summary:
At first, I was trying to fix "DeprecatedDirectThriftFieldAccess error" which is complained in my other diff: D24115197.

And then I realized that we're using the same default queue config from TH3 on TH4. So I decided to correct it with the new values I got from TH4.

But I encountered two questions when I tried to introduce the new values for TH4:
- SharedBytes is actually 0 on TH4 but all the other platforms are using 8 MMU units by default.
- AQM threshold is actually 133168898  rather than 13631418 on both TH3 and TH4.

I created a Broadcom case: CS00011560232 to ask these question to Broadcom. But this diff should be safe to land due to we're not running bcm_test on TH4 yet.

Differential Revision: D25257731

fbshipit-source-id: dd4944f2f050a7f26ebc00ef186f48f498a84afd