Refactor Port/CPU queue stats update
Summary:
In this diff, I've refactored the current update queue stats logic.

Currently, we have both BcmControlPlaneQueueManager and BcmPortQueueManager implement their own updateQueueStat(), which the main logic is the same and the only difference is what exactly the std::pair<bcm_gport_t, bcm_cos_queue_t> we use to call bcm_cosq_stat_get().
So instead of duplicating the code, I moved the main logic back to the parent class: BcmCosQueueManager.
And let getQueueStatIDPair() virtual function to handle the id pair difference.

Also after we merge the updateQueueStat() logic to a single place, we can also include the logic to use BcmEgressQueueFlexCounter to collect the stats
along with the old logic.

Differential Revision: D24148899

fbshipit-source-id: 3f80e01fb588efd2976d26efd9e36cfeab80ae36