Provide a helper api to ensure lock is held
Summary:
In the normal course of things, we do
granular locking during stateChange application.
However, during a transaction rollback, we hold
the lock throughout. So provide a abstraction
that helps stateChanged its downstream functions
ensure that lock is held. If we are in a coarse
lock call chain, ensureLocked is  noop, else
it acquires the lock and passes ownership to
caller

Reviewed By: phshaikh

Differential Revision: D25418089

fbshipit-source-id: c88a749e737188695a4c585299719df73e5c5a2c