PHY refactoring: Added platform mapping to qsfp_service and utility to get PhyPortConfig
Summary:
This checkin puts in the utility to serve the thrift calls from Agent to qsfp_services for the Phy API. The Agent will make thrift call to qsfp_service with port id, port profile and phy id information. Based on that the qsfp_service needs to build PhyPortConfig as per the PlatformMapping present here.

As a part of this checkin:
1. Added platform mapping to qsfp_service so that platform port and phy configuration can be obtained by qsfp service
2. Added a function to generate PhyPortConfig from PlatformMapping. The PhyPortConfig is required by many of ExternalPhy functions

Differential Revision: D23495919

fbshipit-source-id: 5fa15cae4efca8cd6fede9998a675b6fd1c7986b