Add transactional updates in handlePendingUpdates loop
Summary:
As titled. Main goal of the logic is such - transactional updates
need to be executed individually. In order to do this
- If first update in pendingUpdates is transactional splice that and only
that onto the updates list and apply to HW. Rest of the updates will be
handled in the next invocation of handlePendingUpdates
- On encountering a transactional update in the middle of pendingUpdates list
splice all updates upto that transactional update (splice range is [first,
last) - https://en.cppreference.com/w/cpp/container/list/splice). Other updates
will be handled in subsequent invocations of handlePendingUpdates

Note that handlePendingUpdates is called once for each stateUpdate, so
we are guaranteed that as long as pull atleast one update off the
pendingUpdates list we won't need to wait for any more updates before
the current sequence of updates is processed.

Differential Revision: D25388742

fbshipit-source-id: 9bb5435287bf4d10867f01a2482000d3d018cab6