Allow blocking state updates to be transactions
Summary:
As titled. For now only allowing for blocking state
updates to be transactions. Ideally on a failed transaction
we want to throw the error back to to the caller. This
only makes sense for blocking state updates.
However, the throwing back to caller requires readiness
from our clients - bgp, openr etc to be able to act on
this error. This is still being worked out. So for now
I have retained the behavior of retrying failed updates.
But in looking to the future, only allow blocking state
updates to be transactions

Differential Revision: D25391613

fbshipit-source-id: 8bbbf753bcbabae2918fa17fc7873dc71554a14e