Make desired, applied state details private to SwSwitch
Summary:
When desired and applied state diverge, SwSwitch
tries to reconcile them in subsequent SwSwitch updates.
For the external callers, standardize on applieState,
i.e. from the perspective of external observers the
SwSwitch state (SwSwitch::getState) reflects what is
applied in HW. That SwSwitch tries to converge
desiredState == appliedState is a impl detail

Differential Revision: D25392570

fbshipit-source-id: 7a140dacd643e5b380bf431a632452a06aadea0a