Update port json to use port type for physical ports
Summary: A new format for json encoding of port type was added previously for Agg ports. At that time, physical ports were not migrated to use the new format to allow warm boot downgrade to work. Now that the code is deployed, changing physical also to use the new format.

Reviewed By: jasmeetbagga

Differential Revision: D25384140

fbshipit-source-id: 5b2d01a54fc03f294bf85464ff008e074f3a232d