Fix invalid port name crash on galaxy
Summary:
With D25578879, I changed the way we select profiles in PlatformMapping to also use port/pim id. We get pim id using a regex that looks for eth/X/Y. Since this is invalid for galaxy backplane ports, this will fail.
The profile factor only has the pimIDs field if the platform has pims. This change makes it so we only look for pim id on these platforms, therefore not trying to do this on galaxy/rsw platforms.

Reviewed By: jasmeetbagga

Differential Revision: D25593736

fbshipit-source-id: 76efb75926cbd64ccfc07f899d797360e175c6ab