Merge getter of different transceiver info field into one
Summary:
To support CMIS module running in 100G, wedge_agent needs to get the TransceiverManagementInterface from qsfp_service to correctly config the xphy.

With the fact that multiple fields are now needed from TransceiverInfo to help decide the configuration, we removed the usage of getCableInfo and getTransceiverExtendedSpecCompliance and just passing the whole TransceiverInfo to reduce redundency.

Differential Revision: D24916894

fbshipit-source-id: 9a1c4ec1033ccf353879f0f0bd8fdd0a2cfcbe7e