Don't silently absorb error but rather signal error back to updates
Summary:
On encountering a error applying update to HW don't
just absorb the error but rather convey the error back to
updates. The update can then do as it sees fit with this
error. If multiple updates were coelesced together, signal
error to all (since we can't tell which one caused the
error). The updates can then figure out based on the
desired and applied state whether the update was applied
or not. In the common case we will just hard fail on the failed
update. The case where we want to protect against a failed
update will schedule a blocking, non coalescing update
and act on that error appropriately.

TODO:
- Stop queing back failed updates. Now that we signal
error back to the callers, we let them decide what to
do with failed updates
- Get rid of applied, desired state concepts from SwSwitch
and the associated counters. Will follow up if we need
some new counters for conveying failed updates.
- Change route (and maybe neighbor) updates to start
queue updates as blocking and non coelescing and act
on errors appropriately

Differential Revision: D25605665

fbshipit-source-id: 9b5ee694f701aeed27fb0f5a6a1cc5f41c7c7d7a