API to queue updates with Hw failure protection
Summary:
As titled, creating a helper api to queue updates
which desire Hw failure protection. For such updates we if
HwSwitch fails to apply the desired update, we will throw
the error back to the caller and not crash the process.
We will use these types of updates to protect against
FIB overflow and throw the error back to the caller
(viz. bgp, openr)

Differential Revision: D25609318

fbshipit-source-id: 81affa24e708db9c2065d629229c152b78c3b14d