Fix counter related tests
Summary:
Currently copp related tests fails on TH4 because it still uses the old way to get the queue stats
But TH4 needs to use the new FlexCounter implementation to collect queue stats.

In my previous diff, I've already make BcmCosQueueManager::updateQueueStats()
implement both legacy and new FlexCounter queue stat mechanism.
Therefore, we can utilize this function to get the cpu queue stat.

To do that, I decided to pass in BcmEgressQueueTrafficCounterStats* returnQueueStats, which will
be a return map of the queue stat if the map is valid and it has specified what queue and counter type
the caller wants to return.

And after that, I decided to remove BcmTestStatUtils.cpp since it only supports the legacy way to collect queue stats.

Besides, in this diff, I also fixed HwTestQosUtils::disableTTLDecrements() to use routeTable to get the HostRoute
if the asic is not supported HOSTTABLE feature.

Differential Revision: D25683963

fbshipit-source-id: ef6641880e40cdf86110a39afec60d806d6c5798