Ignore hw update errors when agent is exiting
Summary:
As titled. Once agent is signalled to exit, we stop
sending hw updates to HW
https://www.internalfb.com/intern/diffusion/FBS/browse/master/fbcode/fboss/agent/SwSwitch.cpp?commit=e5f744675c6c838f511aebd997881035337018dd&lines=870

Which then causes the update thread loop to mistake this for a
failed (rather than a skipped) update. Reflect this in the update
processing loop.

TODO: See if we can drain the update thread, before transitioning to
EXITING state.

Differential Revision: D25775976

fbshipit-source-id: 932786f5978784bab713bf81a37c2d4fc34439a9