fix profile tests for TH4 devices
Summary:
This diff fixes profile tests for th4 devices. The problem was due to the fact that these tests enable all ports in the first 2 port groups, but only the controlling ports have ingress vlan ids set. This causes an error in the sdk because our default vlan (id 0) is not valid in TH4. We can just set the other ports to the same vlan as the controlling port here.
also make sure we go through bcm port for getting port speed in verifyInterfaceMode since there are different ways to get speed depending on chip type

Differential Revision: D25787763

fbshipit-source-id: 29959b34d19d4058808d0ede5bf99b241cbfb582