Adding the time that Dom data was collected at in wedge_qsfp_util
Summary: Currently, the wedge_qsfp_util doesn't have a timestamp indicating when the dom data was read. This change adds an attribute "timeCollected" to the Sff8636DomData and CmisDomData thrift structs and then uses that in the wedge_qsfp_util client. Ideally, the best place for timeCollected would have been in TransceiverInfo struct but that one doesn't seem to be used by the qsfp_util currently so for the time being this, attribute lives in the 2 dom data structs.

Differential Revision: D25820260

fbshipit-source-id: f546a8e3d0a27387cb2f09213ef7b48e671b03dd