Disable BUFFER_POOL on TH3 for now
Summary:
We don't support BUFFER_POOL on TH3 yet, disable for now.
Enable once the support is added.

For now, the focus is getting one BRCM-SAI test to PASS on TH3.
And we are hitting ASSERTs in SaiBufferManager.
BUFFER_POOL support should not be needed for to get us going with some tests to PASS so disable it for now.

This feature is not used by BcmSwitch either so this should be OK.

Differential Revision: D25908050

fbshipit-source-id: fda4151de3777f91702f5d00df3423c2db579122