Added read/write thrift apis in qsfp_service
Summary: This is the first step in making the functionality in wedge_qsfp_util safer. Currently the wedge_qsfp_util directly talks to the transceiver module for raw register read and write. This change adds a thrift API in qsfp_service which in future will be used by wedge_qsfp_util.

Differential Revision: D25896437

fbshipit-source-id: 6bf88bdbfe700c6504872aeb733fc794cb251b41