Making first optical remediation as 2 mins while keeping regular emediation interval same
Summary:
We see that sometime some of the ports come up quite late. On debugging we found that some of the port (like eth4/15/1 in yamp fa003-uu004.frc2.tfbnw.net) comes up 5 minutes later than all other ports.

After 5 mins the remediateFlakyTransceiver comes up and does recovery on this optics which leads to port bring up. This 5 minute delay is too much and it causes the NHS to fail. In this diff we are making the remediateFlakyTransceiver call just 2 minutes after the startup and then every 5 minutes subsequently.

Differential Revision: D26014510

fbshipit-source-id: c70818773cb86f9d474c43814da509ad7998df89