Unblock TH3 testing on VerifyMacLearningScale
Summary:
Due to T83358080, we found out TH3 takes way longer to learn
MAC compared to other platforms and if the scale is too big, the tests will
be very flaky. For now, change 7K to 5K while waiting for Broadcom to help
us debug this issue.

This diff will change
- the original 7K macs to 5K for scale testing.
- And also remove the original AsicType check in VerifyMacLearningScale, because we should use
known bad tests in the config to manage whether a platform can support a test.
https://fburl.com/diffusion/s24rj4vn

Differential Revision: D25992072

fbshipit-source-id: 321b29026cffedaffd7525330be8d8867ed275df