Code reuse, have clearClassIDs reuse updateClassIDForRoutes helper
Summary:
As titled. This is a edge case function, where we transition
from qph route marking to not. In this case capture all routes to clear
class IDs from using the current state delta's new state (this would be
the first state delta observed after warm boot) and clear classIDs for
routes

Differential Revision: D26118045

fbshipit-source-id: bf56edb44d49fda07c479fdbebbb6da10d97b4e6