return NONE as management interface on exception
Summary:
I have some evidence in testing that we are getting some
garbage values for the TransceiverManagementInterface if we get an
exception when retrieving the value. I didn't think this should happen
since it should be an enum class with a default value of SFF.

However, we want to explicitly return neither SFF or CMIS if we don't
know, so seems reasonable to change logic here, even if we aren't
getting garbage values.

Differential Revision: D26327743

fbshipit-source-id: d9f202ea0f0391976b715ad3523e78ce5c49655e