Allow configuring to cpu routes over thrift
Summary:
As titled. In the backend we support a forwarding action
of drop, cpu, nhops for each route -
https://github.com/facebook/fboss/blob/master/fboss/agent/state/RouteNextHopEntry.h#L70-L90
However over thrift we only allowed for nhop and drop actions.
This is unnecesarily restrictive.

Context: the immediate problem I am trying to solve is to be
able to configure TO_CPU routes on standalone RIB. RIB takes
UnicastRoute as input for update and hence the needed to augment.
I could always change the RIB interface and take in RouteNextHopEntries,
but this seems more flexible

Reviewed By: saifhhasan

Differential Revision: D26235756

fbshipit-source-id: 4476f9d5652ebd4a710d0ea70db8e66e4f1ab53c