expose port hw details via thrift
Summary:
This exposes some port hardware details via thrift. I'm
working on a tool where it would be helpful if we could understand the
internal mapping of the port for triaging an issue. Initially was
thinking that we could use the platform mapping, but since the agent
platform config calculation is somewhat complex, it seemed better to
return after the agent has already applied any overrides.

Reviewed By: birdsoup

Differential Revision: D26446273

fbshipit-source-id: 6422c96951ce616b88e9117f23d02c5d379274ec