managed neighbor not a subscriber for port
Summary:
this is done to make managed neighbor simpler and somewhat agnostic to that of trunk vs port choice.

it is also okay to have neighbor not subscribe for port because neighbor subscribes to fdb entry which subscribes to bridge port, which are created with either sailag or saiport object

in addition, put several TODOs.

Reviewed By: jasmeetbagga

Differential Revision: D26479366

fbshipit-source-id: 70344e8c5f82e3af5f9a0f8adc79b8b6a5a1d359