Pull up class id programming into RouteUpdateWrapper abstraction
Summary:
Since we do this from observer as well as from hw tests,
move the logic to a common place.
Note that I did not absorb this into the common addRoute API
as that entails a route resolution, which can be expensive. Classid
changes never require route resolution. So keeping with the existing
legacy rib approach, I kept these apis separate

Reviewed By: shri-khare

Differential Revision: D26455374

fbshipit-source-id: 190c5caafc8a601d7e8580db7e5aea515e00691a